+++ packages/asc-ui/src/components/StepByStepNav/StepByStepNavStyle.ts
@@ -1,7 +1,8 @@
 import styled, { css } from 'styled-components'
 
 import Typography from '../Typography'
-import { breakpoint, themeColor } from '../../utils'
+import { ascDefaultTheme } from '../../theme'
+import { breakpoint as mqBreakpoint, themeColor } from '../../utils'
 
 export interface Step {
   label: string
@@ -12,6 +13,7 @@ export type ItemType = 'checkmark' | 'numeric' | 'none'
 export interface StepByStepNavProps {
   /** 1-based index determining which of the steps is active */
   activeItem?: number
+  breakpoint?: ReturnType<typeof mqBreakpoint>
   className?: string
   itemType?: ItemType
   steps: Array<Step>
@@ -26,34 +28,46 @@ const checkmarkIcon = `<svg viewBox="0 0 32 32" width="14" height="14" fill="whi
 const itemSize = 20
 const activeItemSize = 36
 
+export const transitionBreakpoint = mqBreakpoint(
+  'max-width',
+  'laptop',
+)({ theme: ascDefaultTheme })
+
 const iconStyle = css`
   background-repeat: no-repeat;
   background-position: center;
   background-image: url('data:image/svg+xml,${checkmarkIcon}');
 `
 
-export const OrderdedList = styled.ol`
+export const OrderdedList = styled.ol<{
+  breakpoint?: StepByStepNavProps['breakpoint']
+}>`
   counter-reset: stepByStep;
   list-style: none;
   margin: 0;
   padding: 0;
 
-  @media screen and ${breakpoint('max-width', 'laptop')} {
+  @media screen and ${({ breakpoint }) => breakpoint} {
     display: flex;
     flex-wrap: nowrap;
   }
 `
 
-export const Label = styled(Typography).attrs({ as: 'span' })<{
+export const Label = styled(Typography).attrs({
+  as: 'span',
+})<{
   itemType: StyledProps['itemType']
+  breakpoint?: StepByStepNavProps['breakpoint']
 }>`
   font-weight: 400;
+  line-height: ${itemSize}px;
 
   &::before {
     background-color: ${themeColor('tint', 'level5')};
     border-radius: 50%;
     content: '';
-    display: block;
+    display: flex;
+    flex-direction: column;
     height: ${itemSize}px;
     left: -${itemSize / 2 + 1}px;
     position: absolute;
@@ -66,13 +80,13 @@ export const Label = styled(Typography).attrs({ as: 'span' })<{
         color: white;
         content: counter(stepByStep);
         counter-increment: stepByStep;
-        font-size: 14px;
+        font-size: 12px;
         font-weight: 500;
         line-height: ${itemSize}px;
       `}
   }
 
-  @media screen and ${breakpoint('max-width', 'laptop')} {
+  @media screen and ${({ breakpoint }) => breakpoint} {
     display: inline-block;
     height: ${activeItemSize}px;
     visibility: hidden;
@@ -86,7 +100,7 @@ export const Label = styled(Typography).attrs({ as: 'span' })<{
 `
 
 export const ListItem = styled.li<StyledProps>`
-  ${({ activeItem, itemType, steps, stepsCompleted }) => css`
+  ${({ activeItem, itemType, steps, stepsCompleted, breakpoint }) => css`
     border-left: 2px solid ${themeColor('tint', 'level5')};
     margin-left: ${itemSize}px;
     padding-bottom: ${itemSize * 2}px;
@@ -129,16 +143,7 @@ export const ListItem = styled.li<StyledProps>`
       }
     }
 
-    ${stepsCompleted &&
-    css`
-      ${Label}:before {
-        background-color: ${themeColor('primary', 'main')};
-
-        ${itemType === 'checkmark' && iconStyle}
-      }
-    `}
-
-    @media screen and ${breakpoint('max-width', 'laptop')} {
+    @media screen and ${breakpoint} {
       border: none !important;
       height: ${activeItemSize}px;
       margin: 0;
@@ -160,16 +165,41 @@ export const ListItem = styled.li<StyledProps>`
         );
       }
 
+      &:nth-child(-n + ${activeItem}) {
+        &:not(:nth-last-of-type(${steps.length - (activeItem - 1)})) {
+          background-image: linear-gradient(
+            to bottom,
+            transparent 0%,
+            transparent calc(50% - 1px),
+            ${themeColor('primary', 'main')} calc(50% - 1px),
+            ${themeColor('primary', 'main')} calc(50% + 1px),
+            transparent calc(50% + 1px),
+            transparent 100%
+          );
+        }
+      }
+
       /* Active list item */
-      &:nth-of-type(${activeItem}) ${Label}:before {
-        top: 1px;
+      &:nth-of-type(${activeItem}) ${Label} {
+        &:before {
+          top: 0;
+        }
       }
     }
+
+    ${stepsCompleted &&
+    css`
+      ${Label}:before {
+        background-color: ${themeColor('primary', 'main')};
+
+        ${itemType === 'checkmark' && iconStyle}
+      }
+    `}
   `}
 `
 
 export default styled.div<StyledProps>`
-  @media screen and ${breakpoint('max-width', 'mobileS')} {
+  @media screen and ${mqBreakpoint('max-width', 'mobileS')} {
     display: none;
   }
 `